core: address review — move shim/DDL apply out of engine and compiler
Move pluginCatalogFromCore into a new internal/core/shim package as
PluginCatalog. Move the engine-neutral DDL apply into a new
internal/core/schema package and have the compiler call it directly, so
no clickhouse-specific code remains in the schema loading loop. Push the
ClickHouse type unwrapping into the engine's column conversion so the
core schema apply only sees canonical scalar types. Rename the core
compile path file to parse_core.go. Remove the unit tests in favor of
end-to-end coverage.
