+++ internal/compiler/compile.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/sqlc-dev/sqlc/internal/engine/clickhouse"
+	coreschema "github.com/sqlc-dev/sqlc/internal/core/schema"
 	"github.com/sqlc-dev/sqlc/internal/migrations"
 	"github.com/sqlc-dev/sqlc/internal/multierr"
 	"github.com/sqlc-dev/sqlc/internal/opts"
@@ -58,7 +58,7 @@ func (c *Compiler) parseCatalog(schemas []string) error {
 
 		if c.coreCatalog != nil {
 			for i := range stmts {
-				if err := clickhouse.Apply(c.coreCatalog, stmts[i].Raw); err != nil {
+				if err := coreschema.Apply(c.coreCatalog, stmts[i].Raw); err != nil {
 					merr.Add(filename, contents, stmts[i].Pos(), err)
 					continue
 				}
